P1 Commit 3:
I added my sound file and declared it's variables. I also created variables and conditionals to define the new rocket coming on screen from the old rocket blasting off. This is to ensure only two rockets can possibly be on screen at a given moment. One blasting off as one moves on to the screen.