P1 Commit 5:
I animated the original rocket so it would change into the rocket fire image and travel past the top of the screen once the launch button is pressed. This also triggers the blast off sound. The airborn class is also added so the original rocket won't interfere with the next rocket.